requirements.txt
pyrogram~=1.2.9
TgCrypto
Pillow~=8.3.1